requirements.txt
flask
yfinance
pandas
numpy
gunicorn
psycopg2-binary